fixed all errors

+++ public/src/components/Logout.jsx
@@ -1,27 +1,31 @@
-import React from 'react'
-import { useNavigate } from 'react-router-dom'
-import axios from 'axios'
-import styled from 'styled-components'
-import {BiPowerOff} from "react-icons/bi"
-
-function Logout() {
-    const navigate = useNavigate();
-    const handleClick =async ()=>{
-        localStorage.clear();
-        navigate("/login")
+import React from "react";
+import { useNavigate } from "react-router-dom";
+import { BiPowerOff } from "react-icons/bi";
+import styled from "styled-components";
+import axios from "axios";
+import { logoutRoute } from "../utils/APIRoutes";
+export default function Logout() {
+  const navigate = useNavigate();
+  const handleClick = async () => {
+    const id = await JSON.parse(
+      localStorage.getItem(process.env.REACT_APP_LOCALHOST_KEY)
+    )._id;
+    const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
     }
+  };
   return (
-    <>
-        <Button onClick={handleClick}>
-            <BiPowerOff/>
-        </Button>
-    </>
-  )
+    <Button onClick={handleClick}>
+      <BiPowerOff />
+    </Button>
+  );
 }
 
 const Button = styled.button`
   display: flex;
-  justify-content: space-between;
+  justify-content: center;
   align-items: center;
   padding: 0.5rem;
   border-radius: 0.5rem;
@@ -33,5 +37,3 @@ const Button = styled.button`
     color: #ebe7ff;
   }
 `;
-
-export default Logout
